opengl-cube/requirements.txt
PyOpenGL
pygame
numpy
pyassimp
PyGLM